Clase 1/requirements.txt
colorama==0.4.6
python-decouple==3.8
